DEFECT: --- WinEDDS Export |Application should handle better the following scenarios. (See Description)
http://kcuraflex/Flex/Application/Mask/View.aspx?AppID=1061368&ArtifactID=1099006&ArtifactTypeID=1000002
---

---
Specific Case|Exporting Dynamic Objects times out and returns error for a record that does not contain a File. (See Steps to Reproduce)

http://kcuraflex/Flex/Application/Mask/View.aspx?AppID=1061368&ArtifactID=1099009&ArtifactTypeID=1000002&LayoutArtifactID=1061404
---

---
WinEDDS Import Production Audit| Import with a Non-Identifier field selected as Overlay Identifier|The value in the audit for Overlay Identifer does not return the Non-Identifier field that was selected instead it returns the Identifier field.

http://kcuraflex/Flex/Application/Mask/View.aspx?AppID=1061368&ArtifactID=1099032&ArtifactTypeID=1000002
---