Defect: SB #DAD901
Changing RDC so that 0/0/0000 dates will default to throwing "Invalid Date"
errors (as before), but making it rely on a config file value that, if set
to false, will simply accept the date and set it to NULL.

You might be asking yourself - "Why, Zaid, are we reverting back to behavior
that we intentionally changed many months ago?  We used to just treat 0/0/0000
as a NULL date, and then customers complained, so we set it to throw an "Invalid
Date" exception.  So what gives?  Why are we going back to the old ways?"

All I can do is shrug, because I have no idea either.

--HG--
branch : release-7.5